refactor: Legacy => endpoint로 마이그레이션
