--- api/endpoint_LEGACY/image/index.ts
@@ -1,20 +1,0 @@
-import { axiosInstance } from '@/api';
-
-export const getPresignedUrl = async ({
-  filename,
-  type,
-}: {
-  filename: string;
-  type?: string;
-}): Promise<{ filename: string; signedUrl: string }> => {
-  const { data } = await axiosInstance.request<{ filename: string; signedUrl: string }>({
-    method: 'GET',
-    url: 'api/v1/presigned-url',
-    params: {
-      filename: encodeURIComponent(filename),
-      type,
-    },
-  });
-
-  return data;
-};
